+++ internal/providers/openstack/openstack.go
@@ -61,6 +61,16 @@ func init() {
 		Name:  "brightbox",
 		Fetch: fetchConfig,
 	})
+	// the ovh platform ID just uses the OpenStack provider code
+	platform.Register(platform.Provider{
+		Name:  "ovh",
+		Fetch: fetchConfig,
+	})
+	// the ovh-baremetal platform ID just uses the OpenStack provider code
+	platform.Register(platform.Provider{
+		Name:  "ovh-baremetal",
+		Fetch: fetchConfig,
+	})
 }
 
 func fetchConfig(f *resource.Fetcher) (types.Config, report.Report, error) {