+++ docs/release-notes.md
@@ -13,6 +13,8 @@ nav_order: 9
 
 ### Features
 
+- Add OVH Baremetal provider
+- Add OVH provider
 
 
 ### Changes